pyproject.toml
[project]
name = "vct-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apache-airflow>=3.0.6",
    "apache-airflow-providers-google>=16.0.0",
    "asyncio>=4.0.0",
    "boto3>=1.42.49",
    "deepmerge>=2.0",
    "google-cloud-storage>=3.9.0",
    "httpx>=0.28.1",
    "ipykernel>=6.31.0",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pyspark>=4.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selectolax>=0.4.6",
]
